+++ packages/appinio_social_share/ios/appinio_social_share/Sources/appinio_social_share/AppinioSocialSharePlugin.swift
@@ -6,7 +6,7 @@ import Photos
 
 
 
-public class SwiftAppinioSocialSharePlugin: NSObject, FlutterPlugin, SharingDelegate {
+public class AppinioSocialSharePlugin: NSObject, FlutterPlugin, SharingDelegate {
 
     private let INSTAGRAM_DIRECT:String = "instagram_direct";
     private let INSTAGRAM_STORIES:String = "instagram_stories";
@@ -28,10 +28,10 @@ public class SwiftAppinioSocialSharePlugin: NSObject, FlutterPlugin, SharingDele
     var flutterResult: FlutterResult!
 
 
-    
+
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "appinio_social_share", binaryMessenger: registrar.messenger())
-    let instance = SwiftAppinioSocialSharePlugin()
+    let instance = AppinioSocialSharePlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
 
@@ -90,19 +90,19 @@ public class SwiftAppinioSocialSharePlugin: NSObject, FlutterPlugin, SharingDele
           result(shareUtil.ERROR)
       }
   }
-    
+
     public func sharer(_ sharer: Sharing, didCompleteWithResults results: [String : Any]) {
         flutterResult(shareUtil.SUCCESS)
      }
-     
+
      public func sharer(_ sharer: Sharing, didFailWithError error: Error) {
          flutterResult(shareUtil.ERROR)
      }
-     
+
      public func sharerDidCancel(_ sharer: Sharing) {
          flutterResult(shareUtil.ERROR)
      }
-    
-    
-     
+
+
+
 }